Вроде как все, добавил сразу две нейронки и кнопку переключения
