+++ Neural/css/neuralStylesheet.css
@@ -24,8 +24,9 @@ button:hover {
     background-color: transparent;
     border: none;
     font-size: 0;
-    cursor: none;
+    cursor: auto;
 }
+
 .invisibleButtonCLass:hover {
     background-color: transparent;
 }
