+++ Neural/five/drawMatrixForFive.js
@@ -43,8 +43,8 @@ export function createMatrix(matrixSize, cords) {
                     cords[Number(element.dataset.row)][Number(element.dataset.col)] = 0;
                 }
                 isMouseDown = true;
+                document.getElementById("startButton").click()
             });
-
             element.addEventListener("mouseover", function() {
                 if (isMouseDown) {
                     mouseCounter++;
