+++ Neural/main.html
@@ -53,8 +53,8 @@
 		<p id="trainNumber"></p>
 	</div>
 </section>
-<script type="module" src="fiftyNeural.js"></script>
-<script type="module" src="drawMatrixForFifty.js"></script>
-
+<script type="module" src="fifty/fiftyNeural.js"></script>
+<script type="module" src="fifty/drawMatrixForFifty.js"></script>
+<script type="module" src="main.js"></script>
 </body>
 </html>
